Remove __syncwarp calls.
On the Volta and Turing architectures, threads within a warp can be
scheduled separately.  It appears (through testing on an RTX 2080
(Turing architecture)) that this causes our kernels with __syncwarp
calls to not terminate.

Based on [0], what happens is probably that a warp can have some
threads inside the 'while(!done)' loop, and some threads done with the
loop.  The threads that are done looping are waiting for the
still-looping threads in a __syncthreads call, so they have not
exited.  For __syncwarp the documentation states that "All non-exited
threads named in mask [defaults to 0xffffffff] must execute a
corresponding __syncwarp() with the same mask, or the result is
undefined.", so it seems that what we did resulted in undefined
behaviour, which according to the run meant nontermination.

Let's try to see what happens without the __syncwarp calls.  Maybe
full-out chaos!

[0] https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#synchronization-functions