Look into AEXCH_GLOBAL_CHUNK_COOP bug on Turing compute capability.
On large histograms this kernel ends up doing inter-block cooperation in global
memory.  __threadfence allows for this, but requires that all threads observe
global memory.

The RTX 2080 reported back what seems like a race condition in this case (and a
histogram size/cooperation level of 4096), so maybe this explicit volatile
qualifier will help.